# Conflicts:
#	doc/nanomsg.txt
#	doc/nn_allocmsg.txt
#	doc/nn_freemsg.txt
#	doc/nn_recvmsg.txt
#	doc/nn_sendmsg.txt
#	src/core/global.c
#	src/nn.h
#	src/utils/chunk.c
#	src/utils/chunk.h
#	src/utils/chunkref.c
#	tests/msg.c

+++ src/transports/ws/cws.c
@@ -40,7 +40,6 @@
 #include "../../utils/cont.h"
 #include "../../utils/alloc.h"
 #include "../../utils/fast.h"
-#include "../../utils/int.h"
 #include "../../utils/attr.h"
 
 #include <string.h>
@@ -98,7 +97,7 @@ struct nn_cws {
     struct nn_chunkref remote_host;
     struct nn_chunkref nic;
     int remote_port;
-    int remote_hostname_len;
+    size_t remote_hostname_len;
 
     /*  If a close handshake is performed, this flag signals to not
         begin automatic reconnect retries. */
